DrbDesc: cosmetics: print the QFI as a number
Qfi is a uint8_t strong typedef, so streaming it (even through num())
emits a raw character; the log line showed an invisible control char
where the QFI value belongs.

+++ src/simu5g/common/binder/Binder.cc
@@ -1381,7 +1381,7 @@ DrbId Binder::createOnDemandDrbForQfi(MacNodeId ueNodeId, Qfi qfi)
             DrbId drbId = assignDrbId(ueNodeId, servingNodeId);
             ab.desc.key = DrbKey(NODEID_NONE, drbId);
             ab.desc.lcid = LogicalCid(num(drbId));
-            EV << "Binder::createOnDemandDrbForQfi - QFI " << num(qfi) << " gets on-demand DRB " << drbId
+            EV << "Binder::createOnDemandDrbForQfi - QFI " << (int)num(qfi) << " gets on-demand DRB " << drbId
                << " at UE " << ueModule->getFullPath() << endl;
             pushDrbToRrcs(ab.ueModule, ab.desc);
         }
